Nsc events android 4 169 attend button (#180)
* Added a DTO

Added a

* Added AttendService

also removed my old attempt at it.

* Added IMPL

also added attendeeDto

* Update HttpRoutes.kt

Added route for attend

* Updated attend button

Updated to move this into its own composable. Also added a consent button and an information window to go with the consent button. Also also moved some strings to the strings.xml file for easy editing.

* Updated the DTO

updated the dto to be the correct format for the back end. Also moved more strings to the strings.xml file

* Update EventDetail.kt

* Update EventDetail.kt

Updated with 201 instead of 200

* Minor updates and more strings.xml

* Update EventDetail.kt

Updated

* Update EventDetail.kt